python/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.1
pydantic==2.6.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pywhispercpp>=1.1.0 
anthropic>=0.30.0
openai>=1.0.0
deepgram-sdk>=3.0.0 
httpx>=0.27.0
tqdm>=4.66.0